New Code
I did some research on google for how to make stuff more streamlined and to try and find why my code would not work. It works more now but crashes when I try and search for an ID and when I create a new task it recognizes but doesnt say it.